Added files after AWS independent project
I worked on this project on AWS, but it was not synced with GitHub. These are the files that I created while working on the project. It is the finalized version, where my model can run and learn from the dataset, and with new information it can predict the amount of streams per song in the 2 most popular music streaming apps; Spotify and Apple Music.